Switch to ISPN 8.1 API in ISPN 8.1 plugin
