+++ plugins/infinispan80/src/main/java/org/radargun/service/Infinispan80EmbeddedService.java
@@ -14,8 +14,8 @@ import org.radargun.utils.Utils;
 public class Infinispan80EmbeddedService extends Infinispan70EmbeddedService {
 
    @ProvidesTrait
-   public InfinispanEmbeddedContinuousQuery createContinuousQuery() {
-      return new InfinispanEmbeddedContinuousQuery(this);
+   public Infinispan80EmbeddedContinuousQuery createContinuousQuery() {
+      return new Infinispan80EmbeddedContinuousQuery(this);
    }
 
    @ProvidesTrait
